refactor: move logic to its on lib.ts file and maintain state
